Refactor ExtractHidUsages to an extension method
